Initial commit
Since this is just a webpage that loads from the README.md, I decided to forego using an existing Jekyll theme as that would be overkill. Instead I used SimpleCSS.

Dark mode is built-in, so that's a plus.

The font is Inter because that's the font of the official QC website.